Refactor checkConnection method, rename SessionModel file to Session
